Changed the ROS package searching algorithm to use "rospack".
It allows the use of mesh files from packages in several workspaces.

Also merged the 2 constructors into a single one with single string arg,
and the differentiation between a string URDF arg and filename arg is done
by searching for "<?xml" at the beginning of the arg. If this string matches
at the beginning of the string arg, it is treated as a URDF passed directly
as string, and if it does not match, then the string arg is treated like
a filename.

The LUA function simExtImportUrdf no longer takes a second string argument
with package path.